pyproject.toml
[project]
name = "vina-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Balasundaram Kadirvelu", email = "balasundaram.k@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "bcrypt==4.0.1",
    "cloudinary>=1.44.1",
    "elevenlabs>=2.33.1",
    "fastapi[standard]>=0.128.0",
    "google-genai>=1.62.0",
    "google-generativeai>=0.8.6",
    "jinja2>=3.1.6",
    "litellm>=1.59.7",
    "moviepy>=2.2.1",
    "opik>=1.10.1",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.3.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.31",
    "tenacity>=9.1.2",
]

[project.scripts]
vina-backend = "vina_backend:main"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]
